Made Error page as a common component.
