+++ src/views/farmer-detail-page/farmer-form-preview/FarmerFormPreview.tsx
@@ -1,8 +1,9 @@
+import { useParams } from "react-router-dom";
 import FarmerFormPreviewLeft from "./FarmerFormPreviewLeft";
 import FarmerFormPreviewRight from "./FarmerFormPreviewRight";
 import { useFetch } from "../../../utils/hooks/query";
 import { ENDPOINTS } from "../../../utils/constants";
-import { useNavigate, useParams } from "react-router-dom";
+import ErrorPage from "../../../components/error-page";
 import { S } from "./farmer-form-preview.styled";
 
 const FarmerFormPreview = () => {
@@ -10,7 +11,6 @@ const FarmerFormPreview = () => {
     formatChangeSuccess: isSuccess,
     result: { data: farmersDetailsById },
   } = useFetch(ENDPOINTS.farmerDetails);
-  const navigate = useNavigate();
   const { farmerId } = useParams();
 
   return (
@@ -21,12 +21,7 @@ const FarmerFormPreview = () => {
           <FarmerFormPreviewRight />
         </S.FarmerFormPreviewMainContainer>
       ) : (
-        <S.CertificateNodataContainer>
-          <S.NoDataErrorText>404 Page Not Found</S.NoDataErrorText>
-          <S.NoDataErrorText2 onClick={() => navigate(-1)}>
-            <i>back</i> Go Back
-          </S.NoDataErrorText2>
-        </S.CertificateNodataContainer>
+        <ErrorPage />
       )}
     </>
   );
