+++ src/views/farmer-detail-page/farmer-form-preview/farmer-form-preview.styled.tsx
@@ -198,28 +198,4 @@ export namespace S {
     fontWeight: "500",
     color: theme.palette.text.primary,
   }));
-
-  export const CertificateNodataContainer = styled(Box)(({ theme }) => ({
-    boxSizing: "border-box",
-    background: theme.palette.bg.main,
-    display: "flex",
-    flexDirection: "column",
-    gap: "2rem",
-    alignItems: "center",
-    justifyContent: "flex-start",
-    padding: "5rem",
-    height: "100%",
-    width: "100%",
-    borderRadius: "0.5rem",
-  }));
-  export const NoDataErrorText = styled(Typography)(({ theme }) => ({
-    fontSize: "1.375rem",
-    fontWeight: "600",
-    color: theme.palette.text.black,
-    textAlign: "center",
-    lineHeight: "1.25",
-  }));
-  export const NoDataErrorText2 = styled(NoDataErrorText)(({ theme }) => ({
-    cursor: "pointer",
-  }));
 }
